requirements.txt
PyYAML==3.11
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
colorama==0.3.3
docker-py==1.5.0
psutil==3.2.2
requests==2.8.0
six==1.10.0
websocket-client==0.32.0
wsgiref==0.1.2
